src-tauri/binaries/requirements.txt
# NitroGen Sidecar Dependencies
pyzmq>=25.0.0
Pillow>=10.0.0
